Configuration pour une mesure toutes les 10ms, capteur avec les plus grand range possible.

+++ soft/dev/src/PersonnalDriver/SDCard/spi/sd_spi.h
@@ -16,7 +16,7 @@
 #include "error.h"
 
 #define WAIT_N_MS(n) {											\
-						uint8_t i;								\
+						uint16_t i;								\
 						for(i = 0 ; i < n ; i++ ) WAIT_1MS();	\
 						}
